Only consider closed milestones for version check.
When releasing i3 version 4.x, this milestone should be closed at
the time since the release is done. The release 4.(x+1) should then
be the next open milestone.

This commit fixes the bot for this scenario when reporting the version
to which a user should update if they're using an unsupported version.